requirements.txt
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
colorama==0.4.6
cryptography==41.0.1
idna==3.4
mysql-connector-python==8.0.26
paramiko==2.7.2
protobuf==4.23.3
pycparser==2.21
PyNaCl==1.5.0
python-nmap==0.6.1
requests==2.26.0
shodan==1.26.0
urllib3==1.26.16
XlsxWriter==3.1.2
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
colorama==0.4.6
cryptography==41.0.1
idna==3.4
mysql-connector-python==8.0.26
paramiko==2.7.2
protobuf==4.23.3
pycparser==2.21
PyNaCl==1.5.0
python-nmap==0.6.1
requests==2.26.0
shodan==1.26.0
urllib3==1.26.16
XlsxWriter==3.1.2
tqdm==4.62.3